requirements.txt
appdirs==1.4.4
black==20.8b1
boto3==1.16.47
botocore==1.19.47
click==7.1.2
Jinja2==2.11.2
jmespath==0.10.0
lark-parser==0.11.1
loguru==0.5.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
pyathena==2.1.0
python-dateutil==2.8.1
python-slugify==4.0.1
regex==2020.11.13
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tenacity==6.3.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.2
